rename backend main django files
